Add files via upload
This file downloads and unzips the original data from the study. Right before that, it downloads the data,.table using the library function. Afterwards, it runs the needed code with a data frame solution. It ends up opening a validation file that it creates to visually inspect that the code is working.